% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_if}
\alias{call_if}
\title{Call Function if Condition is Met}
\usage{
call_if(condition, x, FUN, ...)
}
\arguments{
\item{condition}{expression to be evaluated}

\item{x}{first argument to be passed to \code{FUN} or not}

\item{FUN}{function to be called if \code{condition} is met}

\item{\dots}{further arguments passed to \code{FUN}}
}
\description{
Call Function if Condition is Met
}
