% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemeta_validate.R
\name{codemeta_validate}
\alias{codemeta_validate}
\title{codemeta validate}
\usage{
codemeta_validate(codemeta = "codemeta.json", context = NULL)
}
\arguments{
\item{codemeta}{path/filename to a codmeta.json file}

\item{context}{URL (or path or json string) for the codemeta context. Leave as default or use appropriate DOI for the version; see details.}
}
\description{
Perform basic check of JSON-LD expanding and compacting, verifying that the result
}
\details{
by default, validation will use the original context from the import file.
}
\examples{
ex <- system.file("examples/codemeta.json", package="codemetar")
codemeta_validate(ex)

}
