% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codemeta_cran.R
\name{create_codemeta_cran}
\alias{create_codemeta_cran}
\title{Create a codemeta list object from a CRAN Package DESCRIPTION file}
\usage{
create_codemeta_cran(name, version = NULL)
}
\arguments{
\item{name}{(character) Name of CRAN package.}

\item{version}{(character) Version of CRAN package. Defaults to "newest" version.}
}
\value{
(list) A \code{codemeta} list object.
}
\description{
Create a codemeta list object from a CRAN Package DESCRIPTION file
}
\examples{
\dontrun{
# Create codemeta for the newest package version
cm <- create_codemeta_cran("EML")

# Create codemeta for a specific package version
cm <- create_codemeta_cran("EML", "1.0.3")
}

}
