% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_transform}
\alias{crosswalk_transform}
\title{Crosswalk transform}
\usage{
crosswalk_transform(x, crosswalk_context = NULL,
  codemeta_context = getOption("codemeta_context",
  "http://purl.org/codemeta/2.0"))
}
\arguments{
\item{x}{a JSON list or file with data fields to be crosswalked}

\item{crosswalk_context}{Context to be added to x}

\item{codemeta_context}{the address or contents of codemeta context. Leave at default}
}
\value{
a valid codemeta json description.
}
\description{
Perform JSON-LD expansion of input, followed by compaction into the codemeta context
}
