% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{default_context_if_null}
\alias{default_context_if_null}
\title{Return Given Context or Default Context if NULL}
\usage{
default_context_if_null(codemeta_context)
}
\arguments{
\item{codemeta_context}{context to be checked for \code{NULL}}
}
\value{
\describe{
  \item{\code{codemeta_context}}{if \code{codemeta_context} is not \code{NULL},}
  \item{the value of option \code{codemeta_context}}{if thisoption is set,}
  \item{the value returned by \code{codemetar:::url_codemeta_schema()}}{else.}
}
}
\description{
Return Given Context or Default Context if NULL
}
