% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{drop_context}
\alias{drop_context}
\title{Drop Context}
\usage{
drop_context(x, json_output = FALSE)
}
\arguments{
\item{x}{a JSON list (from \code{\link[jsonlite:read_json]{jsonlite::read_json()}} /
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}) or json object (from
\code{\link[jsonlite:toJSON]{jsonlite::toJSON()}})}

\item{json_output}{logical, should output be a json object or a list?}
}
\value{
a list or json object with the "@context" element removed
}
\description{
Drop context element from json list or json string
}
