% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_file}
\alias{example_file}
\title{Get the Path to an Example File Stored in This Package}
\usage{
example_file(...)
}
\arguments{
\item{\dots}{parts of the paths, finally passed to \code{system.file}}
}
\description{
Get the Path to an Example File Stored in This Package
}
