% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_badges.R
\name{extract_badges}
\alias{extract_badges}
\title{Extract all badges from Markdown file}
\usage{
extract_badges(path)
}
\arguments{
\item{path}{Path to Markdown file}
}
\value{
A tibble with for each badge its text, link and link to
its image.
}
\description{
Extract all badges from Markdown file
}
\examples{
\dontrun{
extract_badges(system.file("examples/README_fakepackage.md", package="codemetar"))
}
}
