% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fails}
\alias{fails}
\title{Does the Evaluation of an Expression Fail?}
\usage{
fails(expr, silent = TRUE)
}
\arguments{
\item{expr}{expression to be evaluated within \code{try(\dots)}}

\item{silent}{passed to \code{\link{try}}, see there.}
}
\value{
\code{TRUE} if evaluating \code{expr} failed and \code{FALSE} if
  the evalutation of \code{expr} succeeded.
}
\description{
Does the Evaluation of an Expression Fail?
}
