% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_json_if}
\alias{from_json_if}
\title{Convert from JSON if Condition is Met}
\usage{
from_json_if(condition, x, ...)
}
\arguments{
\item{condition}{expression to be evaluated}

\item{x}{object passed to \code{\link[jsonlite]{fromJSON}} if
\code{condition} is met}

\item{\dots}{further arguments passed to \code{\link[jsonlite]{fromJSON}}}
}
\description{
Convert from JSON if Condition is Met
}
