% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_json_or_character}
\alias{is_json_or_character}
\title{Check for Class "json" or Character}
\usage{
is_json_or_character(x)
}
\arguments{
\item{x}{object to be checked for its class and mode}
}
\value{
\code{TRUE} if \code{x} inherits from "json" or is of mode character,
  otherwise \code{FALSE}
}
\description{
Check for Class "json" or Character
}
