% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{set_context}
\alias{set_context}
\title{set_context}
\usage{
set_context(x, new_context, json_output = FALSE)
}
\arguments{
\item{x}{a JSON list (from read_json / fromJSON) or json object (from toJSON)}

\item{new_context}{new value for the context element}

\item{json_output}{logical, should output be a json object or a list?}
}
\value{
a list or json object with "@context" element set to \code{context}
}
\description{
set context element in json list or json string
}
