% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_json_if}
\alias{to_json_if}
\title{Convert to JSON if Condition is Met}
\usage{
to_json_if(condition, x, ...)
}
\arguments{
\item{condition}{expression to be evaluated}

\item{x}{object to be converted to JSON}

\item{\dots}{further arguments passed to \code{\link[jsonlite]{toJSON}}}
}
\description{
Convert to JSON if Condition is Met
}
