% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codemeta.R
\name{write_codemeta}
\alias{write_codemeta}
\title{write_codemeta}
\usage{
write_codemeta(pkg = ".", path = "codemeta.json", root = ".",
  version = "2", ...)
}
\arguments{
\item{pkg}{package path to package root, or package name, or description file (character), or a codemeta object (list)}

\item{path}{file name of the output, leave at default "codemeta.json"}

\item{root}{if pkg is a codemeta object, optionally give the path to package root. Default guess is current dir.}

\item{version}{schema version}

\item{...}{additional arguments to \code{\link{write_json}}}
}
\value{
writes out the codemeta.json file
}
\description{
write_codemeta
}
\details{
If pkg is a codemeta object, the function will attempt to
 update any fields it can guess (i.e. from the DESRIPTION file),
 overwriting any existing data in that block. In this case, the package
 root directory should be the current working directory.
}
\examples{
write_codemeta("codemetar")
}
