% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{add_context}
\alias{add_context}
\title{Add Context}
\usage{
add_context(x, context, json_output = FALSE)
}
\arguments{
\item{x}{a JSON list (from \code{\link[jsonlite:read_json]{jsonlite::read_json()}} /
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}) or json object (from
\code{\link[jsonlite:toJSON]{jsonlite::toJSON()}})}

\item{context}{context to be added, in same format as x}

\item{json_output}{logical, should output be a json object or a list?}
}
\value{
a list or json object with "@context" element added
}
\description{
Add context element to json list or json string
}
