% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codemeta.R
\name{create_codemeta}
\alias{create_codemeta}
\title{create_codemeta}
\usage{
create_codemeta(pkg = ".", version = "2", ...)
}
\arguments{
\item{pkg}{package description file, can be path or package name (character), or a codemeta object (list)}

\item{version}{schema version}

\item{...}{additional arguments to \code{\link{write_json}}}
}
\value{
a codemeta list object
}
\description{
create a codemeta list object in R for further manipulation
}
\examples{
cm <- create_codemeta("jsonlite")
cm$keywords <- list("metadata", "ropensci")
write_codemeta(cm)
}
