% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codemeta.R
\name{create_codemeta}
\alias{create_codemeta}
\title{create_codemeta}
\usage{
create_codemeta(pkg = ".", root = ".", id = NULL, ...)
}
\arguments{
\item{pkg}{package path to package root, or package name, or description file (character), or a codemeta object (list)}

\item{root}{if pkg is a codemeta object, optionally give the path to package root. Default guess is current dir.}

\item{id}{identifier for the package, e.g. a DOI (or other resolvable URL)}

\item{...}{additional arguments to \code{\link{write_json}}}
}
\value{
a codemeta list object
}
\description{
create a codemeta list object in R for further manipulation
}
\examples{
cm <- create_codemeta("jsonlite")
cm$keywords <- list("metadata", "ropensci")
write_codemeta(cm)
}
