% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk a column to return a JSON-LD context for it based on codemeta}
\usage{
crosswalk(column, cw_table = NULL, to_json = FALSE)
}
\arguments{
\item{column}{column name from the CrossWalk table (character vector)}

\item{cw_table}{A CrossWalk table. call to
readr::read_csv("https://github.com/codemeta/codemeta/raw/master/crosswalk.csv")
 or use default cached copy.}

\item{to_json}{logical, should the result be turned into a json string?
Default FALSE, which permits easier manipulation such as adding this
context to an existing list before serializing to JSON}
}
\value{
a context list, which can be converted to a json string
}
\description{
Crosswalk a column to return a JSON-LD context for it based on codemeta
}
\examples{
crosswalk("Zenodo")
}
