% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_transform}
\alias{crosswalk_transform}
\title{Crosswalk transform}
\usage{
crosswalk_transform(x, crosswalk_context = NULL,
  codemeta_context = NULL)
}
\arguments{
\item{x}{a JSON list or file with data fields to be crosswalked}

\item{crosswalk_context}{Context to be added to x}

\item{codemeta_context}{the address or contents of codemeta context. If not
specified, the default "https://doi.org/10.5063/schema/codemeta-2.0" is
used. The default is taken from the option "codemeta_context" and can thus
be overridden by setting this option with \code{options(codemeta_context =
<your_context_url>)}.}
}
\value{
a valid codemeta json description.
}
\description{
Perform JSON-LD expansion of input, followed by compaction into the codemeta
context
}
