% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codemeta.R
\name{write_codemeta}
\alias{write_codemeta}
\title{write_codemeta}
\usage{
write_codemeta(pkg = ".", cm = NULL, path = "codemeta.json",
  pretty = TRUE, auto_unbox = TRUE, version = "2", ...)
}
\arguments{
\item{pkg}{package description file, can be path or package name.}

\item{cm}{a codemeta list object, e.g. from \code{\link{create_codemeta}}}

\item{path}{file name of the output, leave at default "codemeta.json"}

\item{pretty}{formatting option to \code{\link{write_json}}}

\item{auto_unbox}{formatting  option to \code{\link{write_json}}}

\item{version}{schema version}

\item{...}{additional arguments to \code{\link{write_json}}}
}
\value{
writes out the codemeta.json file
}
\description{
write_codemeta
}
\examples{
write_codemeta("codemetar")
}
