% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codemeta.R
\name{write_codemeta}
\alias{write_codemeta}
\title{write_codemeta}
\usage{
write_codemeta(pkg = ".", path = "codemeta.json", root = ".",
  id = NULL, use_filesize = FALSE,
  force_update = getOption("codemeta_force_update", TRUE),
  use_git_hook = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{pkg}{package path to package root, or package name, or description file
(character), or a codemeta object (list)}

\item{path}{file name of the output, leave at default "codemeta.json"}

\item{root}{if pkg is a codemeta object, optionally give the path to package
root. Default guess is current dir.}

\item{id}{identifier for the package, e.g. a DOI (or other resolvable URL)}

\item{use_filesize}{whether to try to estimating and adding a filesize by using
\code{base::files.ize()}. Files in \code{.Rbuildignore} are ignored.}

\item{force_update}{Update guessed fields even if they are defined in an
existing codemeta.json file}

\item{use_git_hook}{Whether to create a pre-commit hook requiring
codemeta.json to be updated when DESCRIPTION is changed. Default is \code{FALSE} to avoid
an unwanted alteration of the user's git environment.}

\item{verbose}{Whether to print messages indicating opinions e.g. when
DESCRIPTION has no URL. See \code{\link{give_opinions}}.}

\item{...}{additional arguments to \code{\link{write_json}}}
}
\value{
writes out the codemeta.json file
}
\description{
write out a codemeta.json file for a given package.  This function is
basically a wrapper around create_codemeta() to both create the codemeta
object and write it out to a JSON-LD-formatted file in one command. It can
also be used simply to write out to JSON-LD any existing object created with
`create_codemeta()`.
}
\details{
\strong{Why bother creating a codemeta.json for your package?} R packages
encode lots of metadata in the \code{DESCRIPTION} file, \code{README}, and other
places, telling users and developers about the package purpose, authors,
license, dependencies, and other information that facilitates discovery,
adoption, and credit for your software. Unfortunately, because each
software language records this metadata in a different format, that
information is hard for search engines, software repositories, and other
developers to find and integrate.

By generating a \code{codemeta.json} file, you turn your metadata into a
format that can \href{https://docs.ropensci.org/codemetar/crosswalk}{easily crosswalk} between
metadata in many other software languages. CodeMeta is built on
\href{https://schema.org}{schema.org} a simple \href{https://developers.google.com/search/docs/guides/intro-structured-data}{structured data}
format developed by major search engines like Google and Bing to improve
discoverability in search. CodeMeta is also understood by significant
software archiving efforts such as \href{https://www.softwareheritage.org/}{Software Heritage} Project, which seeks to
permanently archive all open source software.

For more general information about the CodeMeta Project for defining
software metadata, see \url{https://codemeta.github.io}. In particular, new
users might want to start with the \href{https://codemeta.github.io/user-guide/}{User Guide}, while those looking to
learn more about JSON-LD and consuming existing codemeta files should
see the \href{https://codemeta.github.io/developer-guide/}{Developer Guide}.
}
\section{Technical details}{

 If pkg is a codemeta object, the function will attempt to update any
  fields it can guess (i.e. from the DESCRIPTION file), overwriting any
  existing data in that block. In this case, the package root directory
  should be the current working directory.

When creating and writing a codemeta.json for the first time, the function
adds "codemeta.json" to .Rbuildignore and, if the project uses git, adds a
pre-commit hook ensuring that if DESCRIPTION changes, the codemeta.json will
be updated as well unless the DESCRIPTION change is committed with 'git
commit --no-verify'.
}

\examples{
\donttest{
write_codemeta("codemetar", path = "example_codemetar_codemeta.json")
}
}
