% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codemeta.R
\name{write_codemeta}
\alias{write_codemeta}
\title{write_codemeta}
\usage{
write_codemeta(pkg = ".", id = uuid::UUIDgenerate(),
  path = "codemeta.json", pretty = TRUE, auto_unbox = TRUE, ...)
}
\arguments{
\item{pkg}{package description file, can be path or package name.}

\item{id}{an id for the package, such as the Zenodo DOI. a UUID will be generated if none is provided}

\item{path}{file name of the output, leave at default "codemeta.json"}

\item{pretty}{formatting option to \code{\link{write_json}}}

\item{auto_unbox}{formatting  option to \code{\link{write_json}}}

\item{...}{additional arguments to \code{\link{write_json}}}
}
\value{
writes out the codemeta.json file
}
\description{
write_codemeta
}
\examples{
write_codemeta("codemetar")
}
